finished seperating prune.tsx into 4 seperate components with use of context api
